feat: o2-4 Global scope settings
